edit / adminService의 메서드 이동
메서드를 consultantService로 이동하고 그에 맞게 controller 호출 서비스 변경
